erlang: get rid of jsone
Summary:
This removes the dependency of the buck2 erlang toolchain on jsone and migrates to using the vendored json module from OTP 27 (which we'll be able to remove once we migrate to OTP 27).

This is done through some "hacks", notably we always add the path to utility modules (which includes `json`) to test dependencies as well as force-bundle it with the test runner escript (through a magic tag) and the test shell.

Adding it in two places is necessary - the shell/runner use json internally to load the test info dependency list, so they can't have access to it through that.
But we still need to include it in the test info dependency list since that is used on the node running the test itself.

Reviewed By: TheGeorge

Differential Revision: D59054624

fbshipit-source-id: ac777d3b12dc705099641fc7909fcbaf2fef38ab