Add label to opt-in genrule from project root
Summary: We want to run LLVM lit unit tests as part of the LLVM buckification effort. We're using `sh_test` as the entrypoint for tests and args is always running from the project root. This means the configuration on other genrules that are used need to be ran from the project root as well. Create a label for that use case here.

Reviewed By: JakobDegen

Differential Revision: D59311948

fbshipit-source-id: f019809b774b0391996d1ae9b3862130cb981fe4