Heap::alloc_typed_unchecked
Summary:
`ValueOfUnchecked<T>` is meant to be store a type which unpacks to `T`.

So if we allocate a value of type `T` it is not guaranteed it unpacks to `T`, but this is how code is usually written.

But this is safer than calling `alloc` followed by `ValueOfUnchecked::new`.

Used in the following diff D59378142 (and hopefully more).

Reviewed By: JakobDegen

Differential Revision: D59378141

fbshipit-source-id: 05e54aba1d480e412d0f5a1b95f1c03b23a7e218