Add -coverage-prefix-map to swiftc
Summary:
Binaries built with code coverage enabled contains the source files paths in its metadata. By default these paths are absolute paths and it causes 2 problems:
* we need [logic to remove the path prefixes](https://www.internalfb.com/code/fbsource/xplat/test_frameworks/coverage/common/llvm-coverage/src/clean_path.rs) further down in the pipeline.
* coverage builds are not truly reproducible as the same repo revision will produce different binaries depending on where the fbsource checkout is located. (each user has a different home folder. each RE worker has a different working directory...)

This diff adds `-coverage-prefix-map` mapping the the current path to `.` to strip away the location of the repo checkout in the machine running the build. this flag is added to all `swiftc` (it has no effect on builds without coverage)

Reviewed By: milend

Differential Revision: D59354230

fbshipit-source-id: 52a174fbcd6100e722863266cd0983700bb6758e