configs: Do config flag parsing later
Summary:
We currently have a stup where we parse configs flags early on in the daemon logic and then pass those parsed flags through to the main config creation logic. There's no real reason to do that though, so this diff makes us instead pass the unparsed configs through.

One reason this was done was in the interest of D47832942 . However, that never landed and is not even necessary anymore. But also, I disagree with the approach from that diff, I'd rather see us do that logic after all config parsing, and not only apply it to command line flags

Reviewed By: stepancheg

Differential Revision: D59664966

fbshipit-source-id: c7cb479a54f0a792e932730c002b88572c927fb6