configs: Don't absolutize all project paths
Summary:
buckconfig loading supports loading both buckconfigs that are within the project, as well as buckconfigs from outside the project (eg in `/etc/buckconfig.d/`). Currently, we handle this by always absolutizing paths as early as we can, and passing `AbsNormPathBuf`s around. That's really messy and leads to a lot of unnecessary complexity in the code. It also blocks improvements later on in this stack. Instead, let's actually track which are which and handle them separately.

It's easy to tell that this version of the code is better because there are a number of `.relativize` calls that go away. The next few diffs also allow us to pass the project root to fewer places and avoid some canonicalizations, and hopefully simplify tests a bit. All of that is good.

While this diff is big, it's mostly pretty mechanical. The primary thing it does is replace this type:

https://www.internalfb.com/code/fbsource/[8d7e8bf253ff284f63e65b2525d3c7477beb6c14]/fbcode/buck2/app/buck2_common/src/legacy_configs/configs.rs?lines=54-59

With:

```
pub(crate) enum ConfigPath {
    Project(ProjectRelativePathBuf),
    Global(AbsNormPathBuf),
}
```

All the rest just fixes compiler errors

Interestingly, this is a small breaking change. Today, it is possible for configs inside the project to include configs outside the project using relative paths. That's pretty broken behavior though, and we shouldn't support it, so seems fine. I'll write a post

Reviewed By: stepancheg

Differential Revision: D59664984

fbshipit-source-id: e32870da493fbee26768300f768a6a759dd87f30