configs: Non-normalized absolute paths
Summary:
None of this logic fundamentally requires us to use normalized absolute paths, so it's simpler not to.

This is a breaking change, since it means that we do physical instead of logical symlink resolution for includes in non-project buckconfig files.

This will allow me to fix up and then land D59541900 . Maintaining code is hard.

Reviewed By: stepancheg

Differential Revision: D59664965

fbshipit-source-id: 3cf4585ef7ee8b98b8968c320e5ec05aaf265b03