cells: More `root_cell_cell_alias_resolver`
Summary: Eventual goal is to take the cell alias resolvers out of the cell resolver and put them onto dice. When I do that though, I intend to keep the root cell alias resolver around because it's more difficult to remove and less expensive to keep. So explicitly using this function wherever possible is good.

Reviewed By: stepancheg

Differential Revision: D59669545

fbshipit-source-id: d7812a2893e4506db28f640692d1280bf4bda9ca