configs: Parse cwd immediate alias resolvers explicitly
Summary:
As per the previous diff, the end state needs to be that we parse these alias resolvers from disk when we need them, as they will no longer be stored in the cell resolver. So do that.

Unfortuantely, I'm at the point in the stack where trying to do easy-to-understand refactorings is getting hard, so instead I'm adding new code. This is a little bit of copy and paste and a little bit of a performance regression, but that will all be recovered later in the stack.

Reviewed By: stepancheg

Differential Revision: D59725049

fbshipit-source-id: d56761c63c8d80c1055f207f53b8dd25258774b1