Populate linker.command subtarget for dthin-lto builds
Summary:
For local thin-lto builds, the linker.command subtarget on an apple_(library|binary|bundle) produces information
describing the command used to perform the link action for the binaries defined within that target. Prior
to this diff, no information is provided for a distributed thin-lto link. This diff adds support for distributed
thin-lto. The information provided by this subtarget is not the same for local thin-lto builds vs distributed
thin-lto builds, but it is useful information nonetheless.

I plan to use this to write tests that compare flags for dthin-lto vs local thin-lto builds.

Reviewed By: rmaz

Differential Revision: D59840767

fbshipit-source-id: 9d67e823f4deaa718c7cfe1cd158d1fa4fa2ff59