Rename the `root` cell
This commit renames the `root` cell to facilitate embedding it as a dependency to Rust projects.
While the Buck2 repository does not have an official public API, depending on it is the most
reasonable way of writing custom test runners, which is important for OSS users until TestInfo V2
lands.

`root` is a common name for root cells, it is generic and it is the default when running `buck2
init`, making it a bad name that is not descriptive and likely to conflict with downstream cells.

As described in the added docs, this commit is not an ideal solution, since it requires
configuration hacks:
- The root `.buckconfig` must point to Buck2's shim directory.
- Buck2's `.buckconfig` must be edited to remove cells that are already defined in the root.