configs: Explicitly parse all cell configs at daemon startup
Summary:
At daemon startup, we need all cell configs so that we can parse the ignore specs. Soon, `BuckConfigBasedCells` won't include this data anymore, so prepare for that by parsing all the cells explicitly.

This diff results in some duplicate work, more changes are needed until it can be deduplicated

Lots of helper functions which are used in tests next diff

Reviewed By: stepancheg

Differential Revision: D59729565

fbshipit-source-id: 13c85a1ea39ee27fd9b49fe9866ad3711adff3f5