config diffs: Make dice friendly
Summary:
Report the config diffs from within dice, instead of doing it outside of dice at the beginning of the command. This makes it compatible with buckconfigs-on-dice. Details explained in the code

Doing this requires us to bring back the `has_new_configs` field, but with slightly different semantics than before

Reviewed By: stepancheg

Differential Revision: D59727035

fbshipit-source-id: bf0dc6439fa684411fa5593f129c21e242055031