configs: Move buckconfigs into dice
Summary:
It was a long road, but we're finally there :)

This diff moves the majority of buckconfig loading into dice, instead of injecting the loaded buckconfigs. Concretely, `BuckConfigBasedCells` loses its `configs_by_name` field, which was mostly unused by this point.

The performance benefits of this are mostly but not fully realized by this diff, there are still a couple of changes that need to be made

This and the previous diff are a mild breaking change for `--reuse-current-config`. So far, that flag has reused all configs. Instead, it now only reuses those parts of the configs that come from outside the repo (either files or flags), but continues to respect changes in the repo. This was agreed on in the core sync some time ago, so not making a post

Reviewed By: stepancheg

Differential Revision: D59729566

fbshipit-source-id: 0ad682d9e6aebb0cc739ae4f52a340940a3518e4