configs: Parse cell in config args in client
Summary:
This is easier than I thought it would be

Currently, we send the values of `--config` and `--config-file` from the client to the daemon unparsed. However, this is a bit silly.  Both args can contain a reference to a cell, so now the daemon has to do an extra read of the configs just to be able to understand what that means, when the client already could do that.

Instead, we now resolve the cell alias to a `CellRootPathBuf` in the client and send that to the daemon. Besides being much simpler, this should also be a light speedup

Reviewed By: stepancheg

Differential Revision: D59758646

fbshipit-source-id: a9a5725bf59d4aae697658da21fb3e5eabf5d3c2