configs: Don't follow includes when loading cell resolver
Summary:
At the beginning of each command in the daemon, we parse the root cell's config to determine the cell resolver and get a couple of configuration parameters. We currently do a "deep" parse for this, ie following includes. This diff changes that to be a shallow parse, ie to not follow includes. The motivation is that a shallow parse is significantly faster

While this is a mild breaking change, I claim that it does not need a post because 1) for the cell resolver, we do this already in other places, and 2) the remaining configuration paramters generally don't affect build semantics and are just "daemon configuration".

Reviewed By: stepancheg

Differential Revision: D59784630

fbshipit-source-id: f54482ca80faa5bc44f3c415caa2f50072f3b268