server: Remove an incorrect lock
Summary:
This diff removes the whole `CellConfigsLoader` struct, but the interesting part is that it removes the lock inside it.

It looks like this is fine because the thing is called just once, but that's actually not the case and is the reason that this diff is a bug fix.

The `DiceUpdater` may, in general, be invoked more than once for each command. That happens in the case of concurrent commands, where we call it once when the command starts to determine whether the command is using the same state as an existing one, and if not, we'll call it a second time after the currently running command has finished. The second invocation is important, since it ensures that we pick up any changes that have happened while blocking.

The use of a `AsyncOnceLock` prevented that later thing from working correctly, meaning we would - somewhat incorrectly - not reload configs when unblocking the second command.

Unfortunately, this is a bit too hard to write a test for, so not doing it

Reviewed By: cjhopman

Differential Revision: D59784632

fbshipit-source-id: d7436f80b4903f629316ad001d8c77cd8299ba6a