genrule add prefer_local labels
Summary:
This diff adds `genrule_prefer_local_labels.bzl`, which is analogous to `genrule_local_labels.bzl` except it sets `prefer_local` instead of `local`.

The motivation is that I've seen genrules that are using labels like `uses_hg` for performance reasons, not because they have to run local only. This prevents the targets from cross building, meaning we have to remove the label and accept the perf hit if we want to unblock the cross build. This diff will allow us to swap out a "force local" label with a "prefer local" label in those cases.

Reviewed By: scottcao

Differential Revision: D59935648

fbshipit-source-id: fdea1397f27b079a44a90aea31efcffd40825327