Add a new approach to computing BUCK_SCRATCH_PATH
Summary:
Instead of using the oddly computed deferred action_key, derive the
scratch path from an artifact bound to the action. This will be no less stable than
the artifact and makes action digests insensitive to changes in the deferred ids
(like action ordering or refactors to how deferred ids are calculated).

A good bit of this is basically backing out D59864942 and changing the name.

Needs to be behind a config because it changes digests.

Reviewed By: JakobDegen

Differential Revision: D59990828

fbshipit-source-id: 64135ba66f2805f6e281f6847ef13d724a490c4a