docs: Follow sidebar structure more
Summary:
Moves a bunch of the remaining pages to follow the sidebar.

I had to rename `benefits.md` to `compared_to_buck1.md` as otherwise it got picked up as an index page

I tried to enforce this in code but simplifying the configuration of the sidebar even more turns out to be a bit hard, so I'm not doing it

To keep from breaking links, I've added redirects. Those are documented as not being testable in dev builds, but I can see them in the build output so probably they're working fine

Reviewed By: iguridi

Differential Revision: D60024231

fbshipit-source-id: b93c8f67668ad6d68e3ddb9c3c7eb749435bed64