Changing type(x) ==(!=) "y" -> (not) isinstance(x, Y) for fbcode
Summary:
# Context
Because some of build is still using buck1, we cannot just change for prelude. Because of this, some of the diffs on the stack {D59985840} are reverted. This diff stack change type(x) ==(!=) "y" -> (not) isinstance(x, Y)  for bxl based on the repro.
Codemode script: P1492217697

# This diff
This diff changes for fbcode.

Reviewed By: JakobDegen

Differential Revision: D60148438

fbshipit-source-id: f1106fdba08b6080859e412b2a821332f71f022b