docs: Dedicated sidebar for APIs
Summary:
Instead of having one mega-sidebar, split it into two. One for the API pages, and one for the other doc pages

This looks a bit cleaner and is a bit easier to use. The mega-sidebar is unfortunately just a bit difficult to navigate, especially as indentation levels increase.

Will ask for some feedback in chat

Reviewed By: iguridi

Differential Revision: D60024910

fbshipit-source-id: 60ea56181e875085df549736cffb60038b7e787a