Use mode/dev-lg instead of mode/dev
Summary:
Enable link groups for mode/dev CI config. See the discussion on D60138427 for full details.

 ---

`linkme` has a requirement that any binaries using `linkme` must be wholly-statically linked. If the binary is at all split into multiple dylibs e.g. with D56779198, linkme breaks and does not find the slice elements located in those external dylibs.

In order to avoid this breakage, we want to enable static linkage for buck2. However, simply enabling `link_style = "static"` on buck2 binaries causes (a) the resulting binary to be too large and causing a linker error, and (b) the link time to be regressed. Instead, using link groups here allows us to keep rust statically linked while most common C++ dependencies are linked into separate link groups.

This is simply using the default configuration, which seems to work fine. It may be possible to further tune this as desired, although I'll leave that to buck2 team.

An alternative approach here would be to replace `linkme` with `inventory`, as that would guarantee better support for the "plugin / registry" feature without requiring static linkage. That I've left to the buck2 team with D60138427 in order to test.

Reviewed By: JakobDegen

Differential Revision: D60179958

fbshipit-source-id: 101c76ded45213ca96e60607a74a576d65687eca