Generate remote dep file entry in cache uploader
Summary:
Computing fingerprints for dep files is expensive, we should only do it if we are going to upload.

This avoids computing them if the exec platform has uploads enabled, not just the action itself, which was causing this to be computed for user builds and contributed to S423672.

As an alternative to moving this call into `CacheUploader` we could add a `can_upload` method to `CacheUploader` but that would not be simpler, and wouldn't allow tracking this in the upload span, added in the next diff.

The reason for using a trait here (`IntoRemoteDepFile`) is that otherwise it would be necessary to move `DepFileBundle` and the majority of dep file code into `buck2_execute`, which doesn't seem worth it.

Reviewed By: scottcao

Differential Revision: D58857886

fbshipit-source-id: 6a8c56b6e9995ff6dc7c9dbc1158ed4c70ced93a