Split dep file upload spans from cache upload spans
Summary:
- To make it more clear what kind of upload is happening, instead of checking reason = 0 or 1
- Make it more clear that the action_digest for a dep file upload span is not the digest of the original action
- Track dep file upload counts separately

This also means no longer logging dep file uploads to `buck2_cache_uploads`, which was an issue since the number of dep file uploads is far higher than local execution uploads and caused the table retention to shrink from ~1 week to <1 day.

If needed individual dep file uploads could be logged to a different table, but having it in the event logs should be enough, along with the upload counts in buck2_builds.

Reviewed By: iguridi

Differential Revision: D58889528

fbshipit-source-id: 0276791caee5ca5b924830567845944024547524