validator api
Summary:
Implementation of an API between validators (user scripts generating validation result artifacts) and Buck2.

Result of running a validator should be a JSON file with the following schema:

```
{
  "version": int,
  "data": dict,
}
```

where `"data"` dict for `"version": 1` should be a:

```
{
  "status": "success" | "failure",
  "message": None | str
}
```

The current schema is a minimal one which I plan to amend based on user feedback.

Reviewed By: stepancheg

Differential Revision: D59725865

fbshipit-source-id: 4dfcb174938e61cccda5082a485e6d110d03e86f