configs: Fix root config loading at daemon startup
Summary:
The VB on Monday broke things.

At daemon startup, we read buckconfigs for two reasons - first, to load a bunch of configuration for the daemon itself from the rootconfig, and second, to load the ignore specs.

The VB on Monday included D59784630 , which caused us to stop following includes when reading the config for the first of those purposes. This meant that specifically the `project.watchman_merge_base` value, which is behind an include, got ignored.

The particular way I noticed is that it caused us to stop logging `branched_from_revision` and `branched_from_global_rev`

Reviewed By: stepancheg

Differential Revision: D60256520

fbshipit-source-id: c22db3ee053923aaf9e8b05e243d93273f5f7fb6