Migrate actions off of the deferred framework
Summary:
This diff migrates actions to the new deferred approach. Actions are the only
thing using the actual "deferred" bit of deferred, and the handling is pretty
straightforward: an action key resolves to either an action or another action
key. The analysis results hold two maps, one for dynamic outputs that map
key->key and then one for key->action. Typically, even a dynamic output will
do a single analysis result lookup and find it's action (because even dynamic
output actions get bound directly to the dynamic_output and then can be found
with two map lookups, but just one dice analysis results lookup).

Context:
The deferred framework is complex and inflexible for what it enables.

Deferred framework enables these two features:
1. defining a key->value computation where the value is a result of general
analysis (rule analysis, dynamic_output, anon_target, bxl)
2. defining (1) such that the analysis that produces the key, does not itself
produce the value but instead binds it to another analysis that will produce
the value (or further bind it).

Some of the unnecessary complexity of the framework:
1. The framework does not distinguish between things that produce analysis
results (and so can hold values) and the things that are values. This makes it
look like the "general analysis" could be anything that implements Deferred
(but in reality only a single implementation of it does so). Only a single
implementation is actually both a value itself and produces values
(DynamicLambda).
2. The outputs of a deferred are split across its actual Output and things in
records in the DeferredRegistry. This makes it complex to understand.
3. Deferred enforces a structured computation where (a subset of) inputs are
declared as DeferredInput, the deferred framework fetches those and puts them
in a DeferredCtx for the computation to then consume.
4. Feature (2) above is actually only used by a single type of deferred
(ActionKey).
5. Due to deferred framework overhead, it supports a "trivial" flag in the key
and then a bunch of structure to support "trivial" deferreds that don't go through
an expensive dice resolution. In fact, everything except for actually deferred
ActionKey could be using the trivial path, but only some do.
6. The "general analysis" results all have different shapes and the lookup of
things has different implementations.
7. The deferred framework does a lot of type erasure. This leads to lots of
different but similar types and is complex to follow data and program flow.

This stack migrates uses off of the deferred framework.

To support feature (1):

1. use deferred's BaseKey (later renamed to DeferredHolderKey) to be the key to general analysis results.
2. add a consistent place to hold these results (RecordedAnalysisValues) that all analysis produces
3. create type-specific keys (like TransitiveSetKey) that are a pair `(DeferredHolderKey, UniquePerDeferredHolderKeyId)`
4. in RecordedAnalysisValues store a map for that analysis's results of each specific type

To support feature (2):
1. the value in RecordedAnalysisValues for an ActionKey (the only thing using
	(2)) is an enum, either the action or another ActionKey.

Reviewed By: stepancheg

Differential Revision: D59939102

fbshipit-source-id: 05f18a71a69fcdcf311f77a43eb7e0df8556a7f8