Split internal and external aliases
Summary:
Split definitions of internal and external aliases for `--modifier`.

Internal aliases are defined in fbsource.
External aliases are still defined in buck2. These are empty for now since all aliases we've defined so far are from `ovr_config//`.

This will put in the groundwork to  OSS users to use aliases as well.

Reviewed By: JakobDegen

Differential Revision: D59902986

fbshipit-source-id: 42bc10409981add447bc0632828c3591f1fb6159