Remove arbitrary concurrency limiter for uplods and replace with a configurable limit that is disabled by default. (#642)
Summary:
The existing limit doesn't really help to achieve the goals it states as it cannot help in conditions where multiple clients are using RE at the same time, which are the common exercise conditions. Since the flag can be used to limit quickly uploads in case of some issues with RBE I opten in for keeping it in some form, by making it a setting, though it likely isn't very useful, so it's disabled by default. Instead we should rely on other control mechanisms like number of connections and TCP or HTTP/2 flow control or, like in this case, explicit errors coming from the RE protocol itself (like `RESOURCE_EXHAUSTED` or others).

Pull Request resolved: https://github.com/facebook/buck2/pull/642

Reviewed By: iguridi

Differential Revision: D59917910

Pulled By: cjhopman

fbshipit-source-id: 1d22940d1d2edee545620c7401f384246425e491