Removing buck2 analysis dependency from buck2 transition
Summary:
This depency was introduced in D58727177, but it slows down the build. Removing it now through use of late binding.

I am creating the late binding, and to have access to some of the arguments, im moving the enum to the late binding file as well.

Additionally, to make the late binding work, i changed the method into a function. One of the arguments I also moved up, to make it available without importing buck2_analysis.

Reviewed By: perehonchuk

Differential Revision: D60173036

fbshipit-source-id: 4eae422d803d1266cbe6b8b68681f2c2968ab16d