Fix broken type in toolchain.bzl
Summary:
See title. In D60309384, RE was enabled for diff-time tests. However that ended up adding a dependency on the toolchain.bzl file, which had types updated in D54263379.

This one line was causing an error in Autodeps due to the fact that Python cannot handle the `|` operator for some union types.

This is a temporary fix while we figure out a more permanent solution.

Got multiple reports on this in the workplace group:

- https://fb.workplace.com/groups/autodeps.users/posts/2698732626961886/?comment_id=2698765646958584

- https://fb.workplace.com/groups/autodeps.users/permalink/2698809960287486/

Reviewed By: stanleyz510

Differential Revision: D60417799

fbshipit-source-id: e50c62b227d288fccd6b0e3d684476357c5af962