Fork dist_lto logic for Darwin
Summary: I need to significantly refactor the dist_lto logic for Darwin, in particular to support pre-merging. It also carries a bunch of complexity we don't need for Darwin. In order to avoid breaking fbcode builds, and reduce complexity in the long term, copy the file to create a darwin specific version of the dist_lto logic. Branch on linker type to call the darwin specific version for darwin builds, and the existing version otherwise.

Reviewed By: christycylee

Differential Revision: D60181468

fbshipit-source-id: 013a797a976afba546e19a845215281570300c04