Account for empty symbol lists
Summary:
`grep` would return a non-zero exit code if the input symbol list was
empty, even though that's not an error condition here. Use sed instead
to avoid this (awk would also work). This will become relevant in a
future diff where I add the ability to limit nm to only defined symbols
(and we might run it on targets with no defined dynamic symbols).

Reviewed By: JakobDegen

Differential Revision: D59612703

fbshipit-source-id: 89dc310ac9d518280f3504c31c8b6c6f551f61cd