Drop CommandExecutionRequest fields before uploading dep files
Summary:
To resolve a regression to peak memory for builds with a large number of cxx cache hits with remote dep file uploading enabled, which contributed in part to S423412.

The cause of the regression is that uploading dep files increases the time spent in the later phase of `execute` which increases the number of actions concurrently holding on to the input paths in the request (`paths` in `CommandExecutionRequest`).
The actions that spend the most time uploading are also the ones holding the most memory which makes this worse.

After this there is still see some regression to peak memory with cache uploads enabled, but it's much smaller.

Reviewed By: cjhopman

Differential Revision: D60249383

fbshipit-source-id: 3984dbb02a82fcf582ac2792ba304a5eac932a67