docs: Remove `--prelude` flag
Summary:
This is unused: https://fburl.com/scuba/buck2_builds/be37ltma

There's a good reason too, which is that it doesn't work. The things we export from the prelude is actually never a rule object and instead always a function shim around one, and so if you try and use this you won't actually get any useful docs.

This is still used in LSP code, which is why I'm leaving it alone for now

Will require a post

Reviewed By: stepancheg

Differential Revision: D60686896

fbshipit-source-id: 325fdb47a2df3df194fca5164d8eb18f1492d889