tests: Flip defaults on `buck_e2e_test`
Summary:
Like it says in the title. The previous diffs ensured that this won't be a behavior change.

Having macros randomly turn tests on seems bad, so let's not do that.

This is a reland of D60726418 , which I had to revert because I realized it just disabled test configs without turning them back on elsewhere

Reviewed By: stepancheg

Differential Revision: D60741766

fbshipit-source-id: adc9800b17361575b559756cb583e30d6f3932a7