Introduce Upstream Borrow and extract _PyGen_yf
Summary:
This diff introduces a new tool for specifying chunks of code to copy from CPython and wrapping them them into a library as part of CinderX. This runs automatically as part of the Buck build. The overall goal is to enable CinderX to make use of internal CPython features with zero changes to the CPython codebase while keeping the implementation's provenance clear. This should allow us to get to a pip-installable CinderX quicker, and make upgrades a little smoother.

In this diff we extract `_PyGen_yf` from `genobject.c` as an example of usage. See `borrowed.template.[c|h]` for the definition of the "borrowed" library for use in CinderX. `borrowed.template.c` is turned into `borrowed.c` by `UpstreamBorrow.py`. The output for a 3.10 build looks like this: P1504040333, and 3.12: P1504041926.

Reviewed By: alexmalyshev

Differential Revision: D60400793

fbshipit-source-id: 9cda0af0e20b186b186edaf024247548e51344ec