docs: Don't put modules into a doc
Summary:
Putting the `DocModule` into a `Doc` is not adding anything and just makes the code harder to follow. This diff has a little bit of copy and paste, more simplifications afterward

This is a mild behavior change in that now the LSP will render the starlark globals together in a virutal file with the buck2 globals. I don't think it matters, but if we really care I can restore that behavior at the end of the stack

Reviewed By: stepancheg

Differential Revision: D60686898

fbshipit-source-id: b56e5fa93800e9d0225124e79439f726c078fccf