docs: `DocModule` instead of `Doc`
Summary:
And pass an optional `ImportPath`.

This simplifies the code quite a bit, but it's also in pursuit of a concrete goal. The docs code currently has a bunch of global assumptions about the file system structure and I'm trying to remove those, so that it's possible to change it

Reviewed By: stepancheg

Differential Revision: D60686890

fbshipit-source-id: 972f759222c24e63e8af424331e05e4dc2ae7502