docs: Add `starlark-builtins` subcommand
Summary:
Having the `starlark` subcommand be able to produce both builtins and normal documentation makes it harder to maintain and isn't really necessary. Splitting this up into two commands makes it easier to use each of those commands the way we want, separately.

For now, this new command does the same-ish thing as the existing one. I want to change that in a future diff to instead have it do something that is a bit more useful.

Reviewed By: stepancheg

Differential Revision: D60686889

fbshipit-source-id: b0ee57599b41fa11d81f9a0b618d67d3f790a5f2