docs: `buck2_cmd_docs_server` crate
Summary:
Next diff also adds a client version of the crate, following the pattern from `buck2_audit` and `buck2_starlark`.

I added the `_cmd_` to the name since I think that makes it a bit more clear what this is. I'll rename `buck2_audit` and `buck2_starlark` to have the `cmd` too, but looking for feedback first

In the process, also late-bind the implementation of the command

Reviewed By: stepancheg

Differential Revision: D60686910

fbshipit-source-id: 840ce5a37995fee63832249aaceb5fdae328cccd