Enable kotlin/java dep file uploads, again
Summary:
These were disabled as an attempted mitigation during S423672 which turned out not to be necessary to resolve the SEV, which was resolved by disabling user dep file downloads.

Unlike cxx dep files, no memory regression was seen while this was enabled, which is expected since the number of input source files to a given kotlin/java action should be much lower, and since the parallelism of kotlin/java builds is much lower.

Reviewed By: JakobDegen

Differential Revision: D60804816

fbshipit-source-id: ec1e1d3a31bacddc3e925b7385918b6024b63c8f