Use buck2_error instead of anyhow for soft_error
Summary:
We should be able to tag soft errors for the same reason we tag normal errors. Tags are not used yet but will be logged to logview later. This would also simplify logging these to buck2_errors, if we choose to do that.

This is used to pass `Deprecation` tags to soft errors in this stack, but seems useful regardless.

Reviewed By: JakobDegen

Differential Revision: D60694255

fbshipit-source-id: 673c100b6e9ecb8f2fabb200b2a55a7de003690e