make get_configuration_analysis_result directly return FrozenProviderCollectionValue
Summary: `get_configuration_analysis_result` is used to resolve the configuration targets used as select keys. It currently accepts a target label and returns an AnalysisResult, but to support unified constraints, it will eventually have to accept provider labels (subtargets). This diff changes the return type of the function in anticipation of this change, and updates users.

Reviewed By: JakobDegen

Differential Revision: D60854526

fbshipit-source-id: ea30b048cad26bfd856498741adc6c2b52f4cc23