Update soft_error definition, add Deprecation tag
Summary:
Makes the documentation for soft_error match how we currently use it in practice, while allowing us to mark errors that do match the original definition with `deprecation=true`.
These can be logged to a separate logview section later on (along with other categories).

Following diff sets it on a subset of existing soft_errors.

Reviewed By: Will-MingLun-Li

Differential Revision: D60696075

fbshipit-source-id: 64e8b5ea242b0fd4dd527c98376fb9a8f54530ea