get_configuration_analysis_result accepts ProvidersLabel
Summary:
With unified constraints, constraint values will be supplied as subtargets, so get_configuration_analysis_result will need to perform analysis on ProvidersLabels instead of TargetLabels. This diff makes that change and updates direct users of the function to pass ProvidersLabel. It also introduces a utility function `ProvidersLabel::parse` that operates in the exact same way as `TargetLabel::parse`.

Supplying PlatformInfo as a subtarget is currently not supported. Execution platform is also not expected to be passed as a sub target.

Reviewed By: JakobDegen

Differential Revision: D60920051

fbshipit-source-id: d6da5823f50226efde5467cbca571cf361d1e511