Handle RE Errors at an unified location
Summary:
Have a unified place early on that allows us to wrap RE errors if necessary. Allows any future downcasts to use this type instead of `REClientError`

This currently prints duplicated information as `decorate_error`. I decided to do that in a separate diff  D60259183 as this diff has a lot already

Reviewed By: JakobDegen

Differential Revision: D60203455

fbshipit-source-id: 243c9c4b38653227f808f8b4c58f53b3dded36bc