Don't validate certs on RE errors (for now)
Summary: There could be a lot of RE failures happening at once (And also action cache misses are considered errors at this point), so this logic is wrong and could cause thousands of certs check to fire. Should either cache the result or move the check elsewhere. Still haven't decided yet so just going to remove this for now

Reviewed By: JakobDegen

Differential Revision: D61035237

fbshipit-source-id: 33028c8acc8ca224161c413d4f42663595afa845