collect validation info transitively during analysis
Summary:
Introduce a new `TransitiveValidations` entity which forms a sparse configured targets subgraph tracking the presence of `ValidationInfo` providers. Build this graph during analysis which is the cheapest option.

Each node in configured targets graph (`AnalysisResult` actually) might:
1) miss the `TransitiveValidations` which means there are no `ValidationInfo` providers in the whole subgraph rooted at this node
2) reuse the `TransitiveValidations` from dependency which means there is only one dependency with `ValidationInfo` providers in its subgraph
3) have a `TransitiveValidations` without `ValidationInfo` provider if it has e.g. 2 dependencies each containing `TransitiveValidations`
4) have a `TransitiveValidations` with `ValidationInfo` provider if the provider was defined for that target node

Basically that allows us not to pay for what we are not using (no excessive querying for `ValidationInfo` providers for each node while traversing the graph), but has some memory cost which is O(configured target graph).

Reviewed By: stepancheg

Differential Revision: D60961202

fbshipit-source-id: 27020211d82dca59e8933388f3a354f04298f3da