Double stack size for asan, again
Summary:
Some tests are still failing with stack exhaustion even after doubling this (D61057823). Root cause seems to be the rust toolchain update (D60317004).

Not clear why this needs to be so high and it's not clear what toolchain change is causing this, but I'm guessing ASAN perf is not a big deal.

Reviewed By: JakobDegen

Differential Revision: D61244061

fbshipit-source-id: a533453115ec45e67d052e8e5e2cb872c8e7bc1a