Allow multiple trampolines in erlang_test() rule
Summary:
# Context
The `erlang_test()` rule allows to set a "trampoline" via the `_trampoline` attribute. This is a binary that will be inserted before the command to run the test. This is a escape-hatch that allows for instrumentation (e.g. using cgroups), etc.

# Problem
The fact that a single trampoline is allowed makes this non-compositional. For example, it's not possible to have a hierarchy of `erlang_test()` macros, each extending the previous one with additional instrumentation

# This diff
We add a `_trampolines` attribute and deprecate `_trampoline`. All the given trampolines will be added  in front of the test command

Reviewed By: robertoaloi

Differential Revision: D61260035

fbshipit-source-id: c47a9aab3e4498b0d5b924ea6e7a43dd80bbeabe