Topo-sort conditional modifiers
Summary:
Modifiers are resolved one constraint key at a time, and buck2 has to identify the ordering to resolve modifiers before applying them.

This diff revives the topo-sort approach to sort conditional modifiers based on deps on their constraint values to identify the order to resolve constraint keys. This was the approach implemented at first in D50206983, but we decided to try a different, simpler approach to maintain a global list of constraint keys to resolve in D50475897. Unfortunately, that approach quickly became unscalable because there were many dimensions that people wanted to condition their modifiers on (OS, CPU architecture, build mode, thin-lto, etc), and it was not possible to keep up maintaining a manual list. This diff revives the topo-sort approach and deletes the globally maintained list of constraint key ordering.

Reviewed By: samkevich

Differential Revision: D61060741

fbshipit-source-id: 96a2f91ac007a45b3ccea321d6b98ecbc2acc954