Use `apple_ipa_package()` for IPA
Summary:
Move .ipa output to be picked up from the "private" `_ipa_package" target.

This allows separate exec platform resolution for the `apple_package()` and `apple_ipa_package()`. This is required because the ipa creation requiring a host-based bundling toolchain which can clash with the compatibility of a custom packager in `apple_package()`.

Reviewed By: manicaesar

Differential Revision: D61264640

fbshipit-source-id: 051c9bd4c064220ba6105a91585b38b586199f27