Check for cfg!(debug_assertions) rather than cfg!(debug)
Summary:
The buck2 built buck2 in dev mode is still crashing with stack overflow today, which is strange because we are already doubling stack size on `cfg!(debug)`.
One possible issue I see is that I cannot find where `cfg!(debug)` is set in dev modes. Switch this to `cfg!(debug_assertions)`, which is documented in https://doc.rust-lang.org/reference/conditional-compilation.html

Reviewed By: JakobDegen

Differential Revision: D61327905

fbshipit-source-id: aac19b5f5d392e1f1c25e67cd6cefab2fe6508e6