Remove cert validity check on daemon connection
Summary: This change has been in for almost 2 months now and didn't get a single hit https://fburl.com/scuba/buck2_builds/8ljtd8j3. If I'm not mistaken, certs are not required for client to make a connection with the daemon. It's more likely that daemon initialization would fail due to certs but even then the connection itself would succeed

Reviewed By: JakobDegen

Differential Revision: D60941721

fbshipit-source-id: 38f5d9153a6ca67c01f0e31e512707dfb8f8f9c5