DynamicValue
Summary:
Rebase of PR https://github.com/facebook/buck2/pull/619 by Andreas Herrmann andreas.herrmann@tweag.io.

The largest difference is that this is implemented on top of new `dynamic_actions` API.

`dynamic_actions` API changes, but it is not used yet.

Some API tweaks are needed and more tests to be done in follow-up.

Reviewed By: JakobDegen

Differential Revision: D60325893

fbshipit-source-id: 321fb682f0ca9bc4983c54c4a59f372292870d65