Append cert failures on build errors
Summary:
We want to check for certs validity on build failure and present it to users so action can be taken. It'd be best to add this at a centralized location to prevent the possibility of the check being called multiple times, so the proposition to be appended to build errors so it will be visible to users and won't overwrite the actual error.

Users actually serious about fixing issues will definitely see the error and our tracking purposes, logview will just use the last error of the list

Reviewed By: JakobDegen

Differential Revision: D61342082

fbshipit-source-id: e4cf5307f3be32cff3d51927a1c58373feb52a18