Setup stability config for m4a
Summary:
This diff use the constrain setting to setup the stability config for m4a.

The changes are:
1. add constrains setting for stability config
2. check constrain value in the compose compiler
3. set the constrain value in the m4a only

Since we only have m4a for the config now, I pass the config directly here. In a latter diff, I will setting up another constrain value for the m4a only.

Reviewed By: navidqar

Differential Revision: D61116399

fbshipit-source-id: ac17c44e94f5a00707e9c2b12c5bb351802aecb2