Don't downcast to detect client broken pipe errors
Summary:
This downcast fails for some commands (audit), so they do not fail silently or return exit code 130. According to Jakob this is because we convert to anyhow to buck2_error and back again.

The previous diff causes the error message in this case to be noisier which makes this more obvious.

Reviewed By: Will-MingLun-Li

Differential Revision: D61509022

fbshipit-source-id: 00e8c2752ba1267f869a7f499308f7fb5fe9e5a6