Add Environment error tier
Summary:
This is intended for errors in the users environment that are not due to direct or obvious inputs to buck.
- External dependencies
- Resource limitations
- Unexpected OS events
- Potentially ambiguous errors

(we may want to further subdivide these categories later, but for now there are not that many of them)

These are useful to distinguish from input errors because:
- We may not be categorizing them perfectly
- Users may or may not have control over them
- They are more likely to cause a SEV if they regress

So we should be more sensitive to these than the majority of input errors, but less sensitive than tier0 errors.

Reviewed By: Will-MingLun-Li

Differential Revision: D61504282

fbshipit-source-id: 1097fe43f34a6f6ac7d8f5c1bb3959e1757c0acf