rust-project: Don't filter -unittest targets
Summary:
This fixes `with_rust_workspace`. The problem is that if we filter out `foo` in favor of `foo-unittest`, then when we pass that to `expand_and_resolve`, it won't resolve the workspaces, because `foo` is in the workspace, but `foo-unittest` is not.

This filtering was originally added in D56494171 to ensure that D56492667 didn't hurt start up time. The latter has since been reverted. Without it, the filtering isn't necessary, because it will happen via the merge.

Reviewed By: Wilfred

Differential Revision: D61543769

fbshipit-source-id: 8c2677e0ac9f2463f88140971ba83e2083a799ac