rust-project: name the binary `buck2-rust-project`
The name `rust-project` is glamorous but far too generic to be meaningful when
talking about it, and it's not immediately clear what it's even referring to
(imagine searching for it on Google). I think `buck2-rust-project` is *much*
better and also gives a template for follow up tools like e.g. `buck2-cxx-
project`

Signed-off-by: Austin Seipp <aseipp@pobox.com>