Fix daemon connection error tagging
Summary:
We are not currently attaching any tags when daemon connection fails, that should be set by parsing stderr, because we attach a new buck2_error as context to an anyhow::Error using `.with_context(|| daemon_connect_error(paths))`, instead of modifying a buck2_error.

Resolving this by making `daemon_connect_error` add tags to an existing buck2_error.

Replacing the other usage of `ConnectError` with `NestedConstraintMismatch` to better match what it is, and the fact that the error message for that is not from stderr.

Also, changing the type of `StatusWithErr` to buck2_error because otherwise converting a buck2_error to anyhow::Error before formatting causes the last context value to be formatted twice (possible bug?). Converting to buck2_error before formatting resolves the duplicate context value and seems more correct.

Reviewed By: Will-MingLun-Li

Differential Revision: D61754469

fbshipit-source-id: acda846b3067b2bc89e70fd396b501a7cebf5331