Don't wait for the Terminate batch script (y/n) prompt when hitting CTRL+C
Summary:
When running a windows command genrule and interrupting it with CTRL+C, it will send the signal to the underlying process but still show **and wait** for an answer to `Terminate batch script (Y/N)?`. Hitting `y` and enter will lead to a cleaner stop. You can also hit CTRL+C again, but that leads to the process (or potentially sub-processes?) to stay around hanging, slowing down your terminal without a clear way to clean it up - other than restarting powershell.

Digging around, it seems like we can't suppress that prompt (unless we patch `cmd.exe`...), but we can skip having to answer it with `<NUL`: https://superuser.com/a/498798

Reviewed By: KapJI

Differential Revision: D61595378

fbshipit-source-id: 45b612b4191b662fc68757244401388e356d42f5