errors: Add category key
Summary:
We should have a single key for grouping errors that roughly categorizes errors by root cause.
The name category key comes from logview. We can use this as the basis of logview MIDs for errors, but it seems useful to have in other error datasets as well.

With only the source location this will aggregate more than we want, later diffs will add more information to disambiguate errors, starting with tags.

Reviewed By: iguridi

Differential Revision: D61695519

fbshipit-source-id: f80e82b0869a43ba82bac191c8ed3404fd73279d