errors: Keep client errors as errors until they are reported
Summary:
Simplifies adding stderr information to `ClientGrpc` tagged errors as errors, instead of ProcessedErrorReport.

Main change is to append stderr information to client connection errors using `context` instead of modifying `error.processed.message` as a string.
This also standardizes the error formatting a bit and removes the "Error is empty?" case since that should never be hit.

`want_stderr` isn't needed since it's only set to true if `ErrorTag::ClientGrpc` is present, which we can now check for directly.

Reviewed By: iguridi

Differential Revision: D61695520

fbshipit-source-id: 2857fb4922e9df3dcd408545892a562a5fdbd5f5