Fix allocative for `ThinBoxSlice`
Summary:
1. Don't report length zero data as unique, it's statically allocated
2. Don't use `.len()` on slices when it should be the size of the data in bytes

After this diff, I'm not a little more convinced that we should delete allocative than I was before. Having something that's only marginally useful is not so bad, but it's probably not great if it has bugs that make it misleading

Also, maybe allocative should panic if inline nodes are bigger than the thing they claim to be a part of?

Reviewed By: stepancheg

Differential Revision: D61755288

fbshipit-source-id: 779f423f356d79066cb2553226e07559cf5266be